requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
huggingface-hub
